refactor: remove project with github worker
