--- Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github.BackgroundWorker/Kysect.Shreks.Integration.Github.BackgroundWorker.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Application\Kysect.Shreks.Application.Abstractions\Kysect.Shreks.Application.Abstractions.csproj" />
-    <ProjectReference Include="..\Kysect.Shreks.Integration.Github\Kysect.Shreks.Integration.Github.csproj" />
-  </ItemGroup>
-
-</Project>
