--- Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github.BackgroundWorker/ServiceCollectionExtensions.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace Kysect.Shreks.Integration.Github.BackgroundWorker;
-
-public static class ServiceCollectionExtensions
-{
-    public static IServiceCollection AddGithubInviteBackgroundService(this IServiceCollection services)
-    {
-        return services.AddHostedService<GithubInvitingWorker>();
-    }
-}
