+++ Source/Kysect.Shreks.sln
@@ -47,8 +47,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Application.C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Playground.Google", "Playground\Kysect.Shreks.Playground.Google\Kysect.Shreks.Playground.Google.csproj", "{CE106075-41CA-42EA-BAE0-4CE490913555}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Integration.Github.BackgroundWorker", "Infrastructure\Integration\Kysect.Shreks.Integration.Github.BackgroundWorker\Kysect.Shreks.Integration.Github.BackgroundWorker.csproj", "{B4338447-F292-46BB-8A9B-4EFDA534541D}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -115,10 +113,6 @@ Global
 		{CE106075-41CA-42EA-BAE0-4CE490913555}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CE106075-41CA-42EA-BAE0-4CE490913555}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CE106075-41CA-42EA-BAE0-4CE490913555}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B4338447-F292-46BB-8A9B-4EFDA534541D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B4338447-F292-46BB-8A9B-4EFDA534541D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B4338447-F292-46BB-8A9B-4EFDA534541D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B4338447-F292-46BB-8A9B-4EFDA534541D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -140,7 +134,6 @@ Global
 		{D0692446-8F47-4E05-B531-2EBB0B22FC4F} = {7A93BF5E-D7A6-4D8A-AC22-9D5AF09C4880}
 		{50DC8AA3-2C5D-4A8D-9F8F-2A16DC7BECE9} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
 		{CE106075-41CA-42EA-BAE0-4CE490913555} = {DC866D57-D41E-455D-BD7D-85F1DD575823}
-		{B4338447-F292-46BB-8A9B-4EFDA534541D} = {D66AA3ED-1CC1-405E-BFD3-7DBA6EDC0E2E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CD5C67DB-70F8-4C47-A3F4-7C94BECE1657}
